+++ assets/javascripts/kitten/index.js
@@ -236,3 +236,4 @@ export { Uploader } from './components/uploaders/uploader'
 
 // Videos
 export { Video } from './components/videos/video'
+export { GifVideo } from './components/videos/gif-video'