+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `GifVideo` component that waits until sources fail then loads `poster`.
 - Feature: Create `CheckList` component.
 
 ## [2.88.2] - 2020-08-27